Split sample view controllres
